requirements.txt
Pillow~=9.4.0
vk-api~=11.9.9
python-dotenv~=1.0.0